pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "document-processor"
version = "0.1.0"
description = "A document processing system with GPT and Claude integration"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["openai", "anthropic", "gpt", "claude", "document-processing"]
authors = [
    { name = "Angelica Marker", email = "software.developer.angelica@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/MarkerAnn/Process-Transcript-to-Notes"
Repository = "https://github.com/MarkerAnn/Process-Transcript-to-Notes.git"
"Bug Tracker" = "https://github.com/MarkerAnn/Process-Transcript-to-Notes/issues"

[tool.hatch.build.targets.wheel]
packages = ["anthropic_api", "openAI_api"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=anthropic_api --cov=openAI_api"
testpaths = [
    "tests",
]